requirements.txt
# Core LLM & Orchestration
langchain>=0.1.0
langchain-community
langchain-core
langchain-openai
ragas
mlflow

# Database & Vectors
psycopg2-binary      # PostgreSQL driver
pgvector             # pgvector Python client
langchain-postgres   # LangChain PGVector store integration

# Data Processing
pandas
numpy

# Web Interface & API
gradio
fastapi
uvicorn

# Utilities
cookiecutter
python-dotenv
pydantic
openai
click
Sphinx
coverage
# Visualization
plotly
pyvis
scikit-learn
networkx
seaborn
matplotlib

# Quality & Development
pre-commit
ruff
